refactor: Move function to another file
